DELTA DEVEL15-windows-smb-vc-refcount-debugging-20080801 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
define DEBUG_SMB_REFCOUNT in smb.h to activate refcount debugging
versions of the functions that log the file and line number as
well as object and refcount.  This is for use in debugging refcount
leaks.

rename GetPacket and GetNCB to smb_GetPacket and smb_GetNCB, etc.,
to enforce the name space.  Do not confuse with similar Win32 API
functions.

restructure some code that refers to smb_vc_t references in order
to clarify where the refcount is dropped.