DELTA windows-callback-shutdown-20080801 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
When the cache manager is shutting down, return an error to all
callback requests.  Otherwise, it is possible for a callback to get
stuck waiting for a lock and when the lock clears to be left with
nothing but invalid memory addresses to crash on.

In the meantime, the file server is waiting for the response.