DELTA DEVEL15-windows-misc-20080806 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Return CM_ERROR_READONLY if the caller wants PRSFS_DELETE and the
volume is readonly.

In cm_CheckNTDelete() do not call cm_SyncOpDone if cm_SyncOp failed

Do not force the use of fake directory data if the user does not
have PRSFS_READ on the directory.  Let the bulkstatus call take
place now that it is actually being used efficiently.

In ApplyV3SearchDirPatches, use cm_SyncOp(CM_SCACHESYNC_GETSTATUS)
to obtain the status info for Freelance entries instead of the
bulkstatus call.

Fix the truncation of 8.3 names in directory search responses.