DELTA windows-build-do-not-mix-pthread-and-lwp-20081008 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
The rx library is built in lwp and pthread forms.  The methods used to
export data symbols are incompatible between the two.  The ubik library
is built with rx headers but only in the lwp form.  As a result it is
not possible to link the ubik libs without also including the lwp rx
lib.  We have pthreaded servers that require linking to the pthreaded
rx and ubik.  To fix the problem all of the macros that assigned values
to variables and were used by ubik have been converted to functions
(but only for Windows.)