DELTA DELTA15-windows-handle-invalid-utf16-names-20081019 AUTHOR asanka@secure-endpoints.com LICENSE MIT FIXES 116641
Windows will deliver to OpenAFS UTF16 strings that are not valid Unicode
and cannot be converted to UTF8.  Return bad file name errors in those
cases.

Make sure all file server strings once converted to UTF16 can be converted
back.  If not, escape the invalid portions of the string so that the
file can be accessed.