DELTA windows-linked-cells-20081103 AUTHOR jaltman@secure-endpoints.com LICENSE MIT FIXES 123571
The Windows code base was implemented ignorant of linked cells as
defined in the src/auth/cellconfig.h struct afsconf_cell.  The code
that made use of the afsconf_cell objects would leak the memory from
the linkedCell field if is non-NULL.

Add cm_SearchCellFileEx which is an extended version of cm_SearchCellFile
capable of returning a linked cell obtained from the CellServDB file.

Update GetCellConfig to populate the linkedCell field of the afsconf_cell
structure.

Modify cm_cell_t to support a linked cell name.

Modify cm_GetCell_Gen() to populate the cm_cell_t linked cell and enforce
that two cells are linked to each other.

Modify cm_GetVolumeByID() and cm_GetVolumeByName() to perform fallback
to the linked cell if the response is CM_ERROR_NOSUCHVOLUME.