DELTA windows-afsd-dir-prefetch-20081226 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
when prefetching the directory buffers do not call cm_FlushFile to
discard locally modified buffers.  Instead just set the cm_scache_t
bufDataVersionLow to the current version.  out of date buffers will
therefore be ignored without requiring that locks be dropped and
reacquired and without searching all of the buffers associated with
the directory.