DELTA DEVEL15-windows-afsd-btree-lookups-20081226 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
when performing a lookup operation it is acceptable for the data version
of the btree to be greater than the data version of the cm_dirOp_t
under which the operation is being performed.  No change to the btree
can take place and the result of the lookup will maintain consistency
without requiring an additional restart of the directory operation.