DELTA DEVEL15-rx-make-rxdebug-optional-20090113 AUTHOR jaltman@your-file-system.com LICENSE BSD
Permit the rx library to be built without the RXDEBUG preprocessor
definition.  When RXDEBUG is not defined all of the debugging functions
either become stubs or log a message indicating that the library was
built without RXDEBUG so that there will be no surprises.  There
are no changes to the export list.  By default RXDEBUG is defined.

Eventually the definition of RXDEBUG can become a configuration option
to the build system.